+++ impl_core/src/main/java/io/opencensus/implcore/stats/MeasureMapImpl.java
@@ -21,11 +21,16 @@ import io.opencensus.stats.Measure.MeasureLong;
 import io.opencensus.stats.MeasureMap;
 import io.opencensus.tags.TagContext;
 import io.opencensus.tags.unsafe.ContextUtils;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 /** Implementation of {@link MeasureMap}. */
 final class MeasureMapImpl extends MeasureMap {
+  private static final Logger logger = Logger.getLogger(MeasureMapImpl.class.getName());
+
   private final StatsManager statsManager;
   private final MeasureMapInternal.Builder builder = MeasureMapInternal.builder();
+  private volatile boolean hasUnsupportedValues;
 
   static MeasureMapImpl create(StatsManager statsManager) {
     return new MeasureMapImpl(statsManager);
@@ -37,12 +42,18 @@ final class MeasureMapImpl extends MeasureMap {
 
   @Override
   public MeasureMapImpl put(MeasureDouble measure, double value) {
+    if (value < 0) {
+      hasUnsupportedValues = true;
+    }
     builder.put(measure, value);
     return this;
   }
 
   @Override
   public MeasureMapImpl put(MeasureLong measure, long value) {
+    if (value < 0) {
+      hasUnsupportedValues = true;
+    }
     builder.put(measure, value);
     return this;
   }
@@ -61,6 +72,11 @@ final class MeasureMapImpl extends MeasureMap {
 
   @Override
   public void record(TagContext tags) {
+    if (hasUnsupportedValues) {
+      // drop all the recorded values
+      logger.log(Level.WARNING, "Dropping values, value to record must be non-negative.");
+      return;
+    }
     statsManager.record(tags, builder.build());
   }
 }